requirements.txt
channels
ctransformers
Django
langchain
pinecone-client
sentence-transformers
pypdf